完善交易流，商品审核通过后金额将发放到用户账户中

+++ src/main/java/com/wangh7/wht/pojo/ItemType.java
@@ -20,6 +20,6 @@ public class ItemType {
 
     String typeCode;
 
-    float typeDiscountBuy;
-    float typeDiscountSell;
+    double typeDiscountBuy;
+    double typeDiscountSell;
 }
